requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
colorama==0.4.6
fastapi==0.112.0
h11==0.14.0
idna==3.7
pydantic==2.8.2
pydantic_core==2.20.1
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.5
cffi==1.17.1
cryptography==43.0.3
dnspython==2.7.0
ecdsa==0.19.0
jose==1.0.0
mongo==0.2.0
pyasn1==0.6.1
pycparser==2.22
pymongo==4.10.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
unicorn==2.1.1
python-dotenv==1.0.1